start work on network object

+++ station.py
@@ -76,43 +76,45 @@ class Station(object):
             print(ionFile)
 
             date = subprocess.check_output('doy2date %s %s'%(day,year),shell=True)
+            if f==files[0]:
+                startDate = date
             start = subprocess.check_output('date2sec %s 00:00:00'%(date[0:10]),shell=True)[0:11]
             end = subprocess.check_output('date2sec %s 23:59:59'%(date[0:10]),shell=True)[0:11]
-            #print('tropNominal.py -m VMF1 -b %s -e %s -stns %s -append -o tdpIn.tdp'%(start,end,self.name))
-            #os.system('tropNominal.py -m VMF1 -b %s -e %s -stns %s -append -o tdpIn.tdp'%(start,end,self.name))
-            #x = fileinput.input(files='Trees/Nick_0.tree',inplace=1)
-            #for line in x:
-                #if 'IONEXFILE ==*' in line:
-                    #line = line.replace("IONEXFILE ==*",'IONEXFILE == "%s"'%(ionFile))
-            replace('Trees/Nick_0.tree',"GLOBAL_EPOCH ==",'GLOBAL_EPOCH == %s'%(start))
-            replace('Trees/Nick_0.tree',"IONEXFILE ==",'IONEXFILE == %s'%(ionFile))
-            if tree is None:
-                os.system('gd2e.py -runType=PPP -rnxFile %s -gdCov -nProcessors=4 -GNSSproducts /home/nvoss/orbits/sideshow.jpl.nasa.gov/pub/JPL_GPS_Products/Final'%(f))
-            else:
-                os.system('gd2e.py -runType=PPP -treeS Trees -rnxFile %s -nProcessors=4 -gdCov -GNSSproducts /home/nvoss/orbits/sideshow.jpl.nasa.gov/pub/JPL_GPS_Products/Final'%(f))
-            os.system('cp smoothFinal.gdcov %s.gdcov'%(date[:-1]))
-            print('sAVED')
-            with open('smoothFinal.gdcov') as fil:
-                content = fil.readlines() #read the files line by line
-                X = content[1]
-                Y = content[2]
-                Z = content[3]
-                #time,x,ux = splitline(X)
-                #time,y,uy = splitline(Y)
-                #time,z,uz = splitline(Z)
-                #time1 = pd.datetime(2000,1,1,11,59,47)
-                #time2 = pd.to_datetime(0)
-                #timedif = time1-time2
-                #time = pd.to_datetime(time,unit='s')+timedif
-                #df = pd.DataFrame([[time,x,y,z,ux,uy,uz]],columns=['Time','X','Y','Z','UX','UY','UZ'])
-                #self.ts.add_data(df)
-                os.system('mkdir %s'%(f.split('/')[7]))
-                os.system('mkdir %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
-                os.system('cp runAgain %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
-                os.system('cp rtgx_Nick_0.tree.err0_0 %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
-                os.system('cp rtgx_Nick_0.tree.log0_0 %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
+            if not os.path.exists('%s.gdcov'%(date[:-1]) and year!='2006'):
+                #print('tropNominal.py -m VMF1 -b %s -e %s -stns %s -append -o tdpIn.tdp'%(start,end,self.name))
+                #os.system('tropNominal.py -m VMF1 -b %s -e %s -stns %s -append -o tdpIn.tdp'%(start,end,self.name))
+                #x = fileinput.input(files='Trees/Nick_0.tree',inplace=1)
+                #for line in x:
+                    #if 'IONEXFILE ==*' in line:
+                        #line = line.replace("IONEXFILE ==*",'IONEXFILE == "%s"'%(ionFile))
+                replace('Trees/Nick_0.tree',"GLOBAL_EPOCH ==",'GLOBAL_EPOCH == %s'%(start))
+                replace('Trees/Nick_0.tree',"IONEXFILE ==",'IONEXFILE == %s'%(ionFile))
+                if tree is None:
+                    os.system('gd2e.py -runType=PPP -rnxFile %s -gdCov -nProcessors=4 -GNSSproducts /home/nvoss/orbits/sideshow.jpl.nasa.gov/pub/JPL_GPS_Products/Final'%(f))
+                    os.system('gd2e.py -runType=PPP -treeS Trees -rnxFile %s -nProcessors=4 -gdCov -GNSSproducts /home/nvoss/orbits/sideshow.jpl.nasa.gov/pub/JPL_GPS_Products/Final'%(f))
+                os.system('cp smoothFinal.gdcov %s.gdcov'%(date[:-1]))
+                print('sAVED')
+                with open('smoothFinal.gdcov') as fil:
+                    content = fil.readlines() #read the files line by line
+                    X = content[1]
+                    Y = content[2]
+                    Z = content[3]
+                    #time,x,ux = splitline(X)
+                    #time,y,uy = splitline(Y)
+                    #time,z,uz = splitline(Z)
+                    #time1 = pd.datetime(2000,1,1,11,59,47)
+                    #time2 = pd.to_datetime(0)
+                    #timedif = time1-time2
+                    #time = pd.to_datetime(time,unit='s')+timedif
+                    #df = pd.DataFrame([[time,x,y,z,ux,uy,uz]],columns=['Time','X','Y','Z','UX','UY','UZ'])
+                    #self.ts.add_data(df)
+                    os.system('mkdir %s'%(f.split('/')[7]))
+                    os.system('mkdir %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
+                    os.system('cp runAgain %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
+                    os.system('cp rtgx_Nick_0.tree.err0_0 %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
+                    os.system('cp rtgx_Nick_0.tree.log0_0 %s/%s'%(f.split('/')[7],f.split('/')[-1][4:7]))
 
-
+        os.system('netSeries.py -r %s.gdcov -i *.gdcov'%(startDate[:-1]))
             #append position and uncertainty to TS
 def splitline(line):
   index,sta,time,position,unc = line.split(' ')
